--- opnmf/decomposition/__init__.py
@@ -1,2 +1,0 @@
-from . import opnmf
-from . model import OPNMF
