+++ opnmf/tests/test_model.py
@@ -2,7 +2,7 @@ from numpy.testing import assert_array_almost_equal
 import scipy.io as sio
 import pytest
 from pathlib import Path
-from opnmf.decomposition import opnmf, model
+from opnmf import opnmf, model
 
 
 @pytest.fixture(scope="module")
@@ -20,6 +20,24 @@ def test_model(cwd):
     m_H = opnmf_model.components_
 
     W, H, _ = opnmf.opnmf(data, n_components=6)
+
+    assert_array_almost_equal(m_W, W, decimal=4)
+    assert_array_almost_equal(m_H, H, decimal=4)
+
+
+def test_model_rank_selection(cwd):
+    """Test auto rank selection"""
+    opnmf_model = model.OPNMF(n_components='auto')
+    mc = sio.loadmat(cwd / 'data/faces_data.mat')
+
+    data = mc['data']
+    data = data[:4]
+
+    m_W = opnmf_model.fit_transform(data)
+    m_H = opnmf_model.components_
+
+    # This chooses 2 components
+    W, H, _ = opnmf.opnmf(data, n_components=2)
 
     assert_array_almost_equal(m_W, W, decimal=4)
     assert_array_almost_equal(m_H, H, decimal=4)
